pyproject.toml
[project]
name = "narrative-ai"
version = "0.1.6"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agentql==1.13.0",
    "agno==1.7.10",
    "beautifulsoup4==4.13.4",
    "brave-search",
    "crawl4ai==0.7.0",
    "fastapi>=0.116.1",
    "google-genai==1.26.0",
    "groq>=0.31.0",
    "httpx==0.28.1",
    "jira>=3.10.5",
    "playwright==1.54.0",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "sqlalchemy==2.0.41",
    "streamlit>=1.47.1",
    "uvicorn>=0.35.0",
]

[tool.uv.sources]
brave-search = { path = "brave-api" }
